fix(RSS-ECOMM-2_69): fix types and interfaces paths
