+++ src/components/layout/layout.tsx
@@ -1,8 +1,9 @@
-import Header from '../header/header';
+import { Header } from '@/components/header';
+import type { PropsWithChildren } from '@/utils/interfaces';
 
-import type { LayoutProps } from '@/utils/types';
-
-export default function Layout({ children }: LayoutProps): React.JSX.Element {
+export default function Layout({
+  children,
+}: PropsWithChildren): React.JSX.Element {
   return (
     <div className="p-4">
       <Header />
