+++ src/pages/registration/registration-form-first/registration-form-first.tsx
@@ -6,7 +6,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { cn } from '@/lib/utils';
 import { ROUTES } from '@/utils/constantes';
-import type { Props } from '@/utils/types';
+import type { RegistrationStepProps } from '@/utils/interfaces';
 import {
   validateRegistrationCredentials,
   type RegistrationCredentials,
@@ -16,7 +16,8 @@ export default function RegistrationFormFirst({
   onNext,
   className,
   ...props
-}: React.ComponentPropsWithoutRef<'form'> & Props): React.JSX.Element {
+}: React.ComponentPropsWithoutRef<'form'> &
+  RegistrationStepProps): React.JSX.Element {
   const [formData, setFormData] = useState<RegistrationCredentials>({
     email: '',
     password: '',
