+++ src/pages/registration/registration-form-fourth/registration-form-fourth.tsx
@@ -9,7 +9,7 @@ import { defaultAddressForm, ROUTES } from '@/utils/constantes';
 import type {
   RegistrationAddress,
   RegistrationFormFourthProps,
-} from '@/utils/types';
+} from '@/utils/interfaces';
 import { registrationAddressSchema } from '@/utils/validations';
 
 export default function RegistrationFormFourth({
