+++ src/pages/registration/registration-form-second/registration-form-second.tsx
@@ -8,7 +8,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { cn } from '@/lib/utils';
 import { ROUTES } from '@/utils/constantes';
-import type { Props } from '@/utils/types';
+import type { RegistrationStepProps } from '@/utils/interfaces';
 import {
   validateUserFormData,
   type RegistrationUser,
@@ -18,7 +18,8 @@ export default function RegistrationFormSecond({
   onNext,
   className,
   ...props
-}: React.ComponentPropsWithoutRef<'form'> & Props): React.JSX.Element {
+}: React.ComponentPropsWithoutRef<'form'> &
+  RegistrationStepProps): React.JSX.Element {
   const [formData, setFormData] = useState<RegistrationUser>({
     name: '',
     lastName: '',
