+++ src/utils/interfaces.ts
@@ -1,0 +1,19 @@
+export interface RegistrationAddress {
+  country: string;
+  city: string;
+  street: string;
+  house: string;
+  postalCode: string;
+}
+
+export interface RegistrationStepProps {
+  onNext: () => void;
+}
+
+export interface RegistrationFormFourthProps {
+  className?: string;
+}
+
+export interface PropsWithChildren {
+  children: React.ReactNode;
+}
