+++ src/utils/types.ts
@@ -8,21 +8,3 @@ export type Routes = (typeof ROUTES)[keyof typeof ROUTES];
 export type RegistrationAddressFormData = z.infer<
   typeof registrationAddressSchema
 >;
-
-export interface RegistrationAddress {
-  country: string;
-  city: string;
-  street: string;
-  house: string;
-  postalCode: string;
-}
-export interface RegistrationFormFourthProps {
-  className?: string;
-}
-export type Props = {
-  onNext: () => void;
-};
-
-export interface LayoutProps {
-  children: React.ReactNode;
-}
